Allow incompletely bucketed tables for Hive
Hive 2.x does not create empty files for bucketed tables that do not fill all
of the buckets, whereas Hive 1.x does. This patch allows Presto to handle
tables that only have some of the bucket files created.

It's impractical to test on both Hive 1.x and 2.x, so this commit adds
unit tests covering the incomplete bucketing scenario.